slow vali cycle correct some data fields that were getting mixed up fix bug in tracking already scored videos
