+++ bitcast/validator/reward.py
@@ -22,7 +22,8 @@ import bittensor as bt
 import json
 from google.oauth2.credentials import Credentials
 from bitcast.validator.briefs import get_briefs
-from bitcast.validator.socials.youtube.youtube_scoring import eval_youtube, reset_scored_videos
+from bitcast.validator.socials.youtube.youtube_scoring import eval_youtube
+from bitcast.validator.socials.youtube import youtube_utils
 
 def reward(uid, briefs, response) -> dict:
     """
@@ -70,7 +71,7 @@ def get_rewards(
     yt_stats_list = [reward(uid, briefs, response) for uid, response in zip(uids, responses)]
     scores_matrix = np.array([yt_stats["scores"] for yt_stats in yt_stats_list])
 
-    reset_scored_videos()
+    youtube_utils.reset_scored_videos()
 
     rewards = normalise_scores(scores_matrix)
